仕様変更 / idの型を修正
データベース内で整数を扱った際、long で返却されるので、int では値を扱えないため。

+++ TodoManager2/model/Tag.cs
@@ -21,8 +21,8 @@ namespace TodoManager2.model {
             }
         }
 
-        private int id;
-        public int ID {
+        private long id;
+        public long ID {
             get => id;
             set {
                 id = value;
